Created conservative code and generated results logs
Current code fails both in its finite difference and conservative
versions. To compile with one or the other, the compiler variable
conservative must be defined. Currently the conservative code blows up
shortly after it starts, independent of the initial conditions, while
the finite difference version, although it does not explode, it produces
an enourmously asymmetric behavior between south and north pole, and
complete irregularity in its values, which is unphysical.