Fixed diff. eq. for conservative case, still explodes
The equation used both for the finite difference and conservative
schemes had errors in its derivation. After fixing them on the
conservative case, code still explodes. Problems arise at the center of
the star, and are due to the resistive terms, in particular the ones
expressing flux in theta. Probably a functional dependance for the
resistivity is needed, rather than a constant value.