Several changes, conservative code now seems to work
-	Corrected error in conservative code, sin(2*theta) was used
due to an error in the code, which produced discontinuities in the
equator. It now seems to work (at least it does not explode).
-	Implemented a corrected version for the finite difference
scheme. However, it seems to be very unstable.
-	Updated the way results are logged, everytime "plotSteps" timesteps
pass, the data is logged in a way readable by gnuplot to produce color
maps for B in the entire star.