Included output plotter
Created the script plotter.py, which uses matplotlib to plot the output
of as_toroidal. This script reads the params.dat file in the output and
extracts from there the required info. It's output is a set of pngs,
which can be transformed into a movie using mencoder with the command:

mencoder "mf://*.png" -mf fps=10 -o test.avi -ovc lavc -lavcopts
vcodec=msmpeg4v2:vbitrate=800

independent of the initial values, the MF seems to aquire a skewed
shape, and then collapses along a "line". Not sure if its just a
numerical issue.