Code persists on hall scale, but does only garbage
Using a hybrid approach for the evolution of beta, where the alpha terms
are not conservative, and using a cutoff at the borders, so no strong
discontinuities are produced, code now evolves on a hall timescale for
certain purely poloidal cases. However, Brutal increase in poloidal
energy indicates a severe flaw, and that this does not reflect the
physical system.