Eliminate artificial softening of beta
Eliminated artificial softening of beta, where the field at the point
just below the boundary was solved via three point interpolation in the
radial direction, instead of time evolving it. This was done because of
the impossibility of time evolving alpha just at the surface with the
previous code.

-Time evolve alpha right at the surface. This requires a point just
outside the star to fix the radial derivative there. For beta, a point
just outside is also used, and chosen in such a way that the radial
derivative of beta is solved backwards at the surface.
-Allow the code to run a certain number of time-steps only through Ohm,
so boundary issues are softened. This is very successful for Ohm modes
as initial conditions, but can severely distort equilibrium conditions.
-Alter plotter to ignore the points that evolve only through Ohmic
dissipation.