Added small script to quantify changes in field.
Added the script field_variation.py, which compares the field at a given
time with the fundamental Ohm mode, and the initial condition, after
having normalized the fields to a constant value. This allows us to
quantify changes in sctructure, something that is obscured in an
analysis that only takes the energies into account.